% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Resta dos números}
\usage{
resta(a, b)
}
\arguments{
\item{a}{un número}

\item{b}{un número}
}
\value{
La resta de \code{a} y \code{b}
}
\description{
Resta dos números
}
\examples{
resta(1, 1)

}
