% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma dos números}
\usage{
suma(a, b)
}
\arguments{
\item{a}{un número}

\item{b}{un número}
}
\value{
La suma de \code{a} y \code{b}
}
\description{
Suma dos números
}
\examples{
suma(1, 1)

}
